Добавляет состояния, разбивает проект на компоненты
