--- src/components/add-review/add-review.jsx
@@ -1,84 +1,0 @@
-import React from "react";
-import {Link} from "react-router-dom";
-
-const AddReview = () => {
-  const addRating = () => {
-    const maxRate = 5;
-    const rateBlocks = [];
-
-    for (let i = 1; i <= maxRate; i++) {
-      rateBlocks.push(
-          <React.Fragment key={`rate-${i}`}>
-            <input className="rating__input" id={`star-${i}`} type="radio" name="rating" value={i} defaultChecked={i === 3 ? `checked` : !`checked`} />
-            <label className="rating__label" htmlFor={`star-${i}`}>Rating {i}</label>
-          </React.Fragment>
-      );
-    }
-
-    return rateBlocks;
-  };
-
-  return (
-    <section className="movie-card movie-card--full">
-      <div className="movie-card__header">
-        <div className="movie-card__bg">
-          <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel" />
-        </div>
-
-        <h1 className="visually-hidden">WTW</h1>
-
-        <header className="page-header">
-          <div className="logo">
-            <Link to="/" className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </Link>
-          </div>
-
-          <nav className="breadcrumbs">
-            <ul className="breadcrumbs__list">
-              <li className="breadcrumbs__item">
-                <a href="movie-page.html" className="breadcrumbs__link">The Grand Budapest Hotel</a>
-              </li>
-              <li className="breadcrumbs__item">
-                <a className="breadcrumbs__link">Add review</a>
-              </li>
-            </ul>
-          </nav>
-
-          <div className="user-block">
-            <div className="user-block__avatar">
-              <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-            </div>
-          </div>
-        </header>
-
-        <div className="movie-card__poster movie-card__poster--small">
-          <img src="img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster" width="218" height="327" />
-        </div>
-      </div>
-
-      <div className="add-review">
-        <form action="#" className="add-review__form">
-          <div className="rating">
-            <div className="rating__stars">
-              {addRating()}
-            </div>
-          </div>
-
-          <div className="add-review__text">
-            <textarea className="add-review__textarea" name="review-text" id="review-text" placeholder="Review text"></textarea>
-            <div className="add-review__submit">
-              <button className="add-review__btn" type="submit">Post</button>
-            </div>
-
-          </div>
-        </form>
-      </div>
-
-    </section>
-  );
-};
-
-export default AddReview;
