--- src/components/my-list/my-list.jsx
@@ -1,73 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import {Link} from "react-router-dom";
-
-const MyList = (props) => {
-  const {moviesList} = props;
-
-  return (
-    <div className="user-page">
-      <header className="page-header user-page__head">
-        <div className="logo">
-          <Link to="/" className="logo__link">
-            <span className="logo__letter logo__letter--1">W</span>
-            <span className="logo__letter logo__letter--2">T</span>
-            <span className="logo__letter logo__letter--3">W</span>
-          </Link>
-        </div>
-
-        <h1 className="page-title user-page__title">My list</h1>
-
-        <div className="user-block">
-          <div className="user-block__avatar">
-            <img src="img/avatar.jpg" alt="User avatar" width="63" height="63" />
-          </div>
-        </div>
-      </header>
-
-      <section className="catalog">
-        <h2 className="catalog__title visually-hidden">Catalog</h2>
-
-        <div className="catalog__movies-list">
-          {
-            moviesList.map((movie, index) => (
-              <article key={`${movie}-${index}`} className="small-movie-card catalog__movies-card">
-                <div className="small-movie-card__image">
-                  <img src={`img/` + movie.src} alt={movie.name} width="280" height="175" />
-                </div>
-                <h3 className="small-movie-card__title">
-                  <a className="small-movie-card__link" href="movie-page.html">{movie.name}</a>
-                </h3>
-              </article>
-            ))
-          }
-        </div>
-      </section>
-
-      <footer className="page-footer">
-        <div className="logo">
-          <Link to="/" className="logo__link logo__link--light">
-            <span className="logo__letter logo__letter--1">W</span>
-            <span className="logo__letter logo__letter--2">T</span>
-            <span className="logo__letter logo__letter--3">W</span>
-          </Link>
-        </div>
-
-        <div className="copyright">
-          <p>© 2019 What to watch Ltd.</p>
-        </div>
-      </footer>
-    </div>
-  );
-};
-
-MyList.propTypes = {
-  moviesList: PropTypes.arrayOf(
-      PropTypes.shape({
-        name: PropTypes.string.isRequired,
-        src: PropTypes.string.isRequired,
-      })
-  )
-};
-
-export default MyList;
