+++ src/components/screen-player/screen-player.jsx
@@ -1,11 +1,14 @@
 import React from "react";
+import PropTypes from "prop-types";
 
-const Player = () => {
+const Player = (props) => {
+  const {onExitBtnClick} = props;
+
   return (
     <div className="player">
       <video src="#" className="player__video" poster="img/player-poster.jpg"></video>
 
-      <button type="button" className="player__exit">Exit</button>
+      <button type="button" className="player__exit" onClick={onExitBtnClick}>Exit</button>
 
       <div className="player__controls">
         <div className="player__controls-row">
@@ -38,3 +41,7 @@ const Player = () => {
 };
 
 export default Player;
+
+Player.propTypes = {
+  onExitBtnClick: PropTypes.func.isRequired
+};
