+++ src/mocks/genres.js
@@ -1,0 +1,12 @@
+export default [
+  {name: `All genres`, active: true},
+  {name: `Crime`, active: false},
+  {name: `Comedies`, active: false},
+  {name: `Documentary`, active: false},
+  {name: `Dramas`, active: false},
+  {name: `Horror`, active: false},
+  {name: `Kids & Family`, active: false},
+  {name: `Romance`, active: false},
+  {name: `Sci-Fi`, active: false},
+  {name: `Thrillers`, active: false}
+];
